refactor: Change directory structure about application
