+++ src/main/java/org/likelion/recruit/resource/application/repository/QuestionRepository.java
@@ -1,6 +1,6 @@
-package org.likelion.recruit.resource.question.repository;
+package org.likelion.recruit.resource.application.repository;
 
-import org.likelion.recruit.resource.question.domain.Question;
+import org.likelion.recruit.resource.application.domain.Question;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface QuestionRepository extends JpaRepository<Question, Long> {
